Added custom multiblock interactions
Boiler now has a pilot light, to ignite, right click a torch above the fuel input
Pilot light mode will burn and keep at 20C using 1Mb per tick if only fueled
If has valid water, it will heat up and turn to steam
Redstone on kicks it down to pilot light mode

+++ src/main/java/mctmods/immersivetechnology/common/blocks/multiblocks/recipe/BoilerFuelRecipe.java
@@ -21,10 +21,10 @@ public class BoilerFuelRecipe extends MultiblockRecipe {
 
     public FluidTagInput fuel;
     private final int time;
-    private final int heatPerTick;
+    private final double heatPerTick;
     Lazy<Integer> totalProcessTime;
 
-    public BoilerFuelRecipe(ResourceLocation id, FluidTagInput fuel, int time, int heatPerTick) {
+    public BoilerFuelRecipe(ResourceLocation id, FluidTagInput fuel, int time, double heatPerTick) {
         super(Lazy.of(() -> ItemStack.EMPTY), ITRecipeTypes.BOILER_FUEL, id);
         this.fuel = fuel;
         this.time = time;
@@ -53,5 +53,5 @@ public class BoilerFuelRecipe extends MultiblockRecipe {
     @Override
     public int getMultipleProcessTicks() { return 0; }
 
-    public int getHeatPerTick() { return heatPerTick; }
+    public double getHeatPerTick() { return heatPerTick; }
 }
